Added MyPathIntegrator and NormalIntegrator
MyPathIntegrator has fixed a little bug in PathIntegrator

In PathIntegrator, the "if(bounces>=maxDepth)" is before "if(!isect.bsdf)".
This will cause some bug in a special case:
When the ray hits a optically inactive medium boundary, isect.bsdf will have no bsdf.
Then a new ray is spawned along the same direction, and bounce--;
However, if bounces==maxDepth, the ray will be terminated errorly before the statement "if(!isect.bsdf)"
In this case, the ray is terminated early just because it hits the inactive boundary.
This will cause a loss of some radiance.

MyPathIntegrator defered "if(bounces>=maxDepth)" so that this part of radiance can be accounted correctly.